requirements.txt
hydra-core==1.3.2
invoke==2.2.0
matplotlib==3.10.0
omegaconf==2.3.0
pytest==8.3.4
pytorch_lightning==2.5.0.post0
scikit_learn==1.6.1
torch==2.5.1+cu118
typer==0.15.1
